Credito a livelli 2/4/6 (preventivatore al 2% base), 'Prova automatica immediata' al posto di 'Prova di stampa gratuita', frase unica 'Pronti per la spedizione entro venerdi' 11 settembre' su tutto il sito

+++ src/lib/utils/shipping.ts
@@ -14,5 +14,5 @@ export function estimatedShipDate(businessDays = 5, from = new Date()): Date {
 }
 
 export function formatItDate(d: Date): string {
-	return new Intl.DateTimeFormat('it-IT', { weekday: 'short', day: 'numeric', month: 'long' }).format(d);
+	return new Intl.DateTimeFormat('it-IT', { weekday: 'long', day: 'numeric', month: 'long' }).format(d);
 }
